deps: V8: cherry-pick 1441665e0d87
Original commit message:

    [riscv] Fix the RISC-V build.

    Due to recent changes, there were missing implementations of various
    methods needed in the simulator and Maglev. Additionally, a static_cast
    is needed in the assembler to silence a warning.

    Port commit dfc894cd22d86ce42830e3bfdf485d963f6396ad
    Port commit c33af9bd408eadd6b62571f862bcb5b763c98ad9

    Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6396542
    Commit-Queue: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
    Cr-Commit-Position: refs/heads/main@{#99706}

Refs: https://github.com/v8/v8/commit/1441665e0d87b541932390385101840758b44924
PR-URL: https://github.com/nodejs/node/pull/60989

+++ deps/v8/src/codegen/riscv/assembler-riscv-inl.h
@@ -115,8 +115,9 @@ void Assembler::set_target_compressed_address_at(
     Address pc, Address constant_pool, Tagged_t target,
     WritableJitAllocation* jit_allocation, ICacheFlushMode icache_flush_mode) {
   if (COMPRESS_POINTERS_BOOL) {
-    Assembler::set_uint32_constant_at(pc, constant_pool, target, jit_allocation,
-                                      icache_flush_mode);
+    Assembler::set_uint32_constant_at(pc, constant_pool,
+                                      static_cast<uint32_t>(target),
+                                      jit_allocation, icache_flush_mode);
   } else {
     UNREACHABLE();
   }
